add checks not to run shell scrips multiple times and update README

+++ provision/shell/install_puppet.sh
@@ -28,22 +28,29 @@
 
 # gem install puppet
 
+if [ ! -f tmp/puppet-repo-added ]; then
+    #http://blog.doismellburning.co.uk/2013/01/19/upgrading-puppet-in-vagrant-boxes/
+    apt-get install --yes lsb-release
+    DISTRIB_CODENAME=$(lsb_release --codename --short)
+    DEB="puppetlabs-release-${DISTRIB_CODENAME}.deb"
+    # Assume that this file's existence means we have the Puppet Labs repo added
+    DEB_PROVIDES="/etc/apt/sources.list.d/puppetlabs.list" 
+    
+    if [ ! -e $DEB_PROVIDES ]
+    then
+        # Print statement useful for debugging, but automated runs of this will interpret any output as an error
+        # print "Could not find $DEB_PROVIDES - fetching and installing $DEB"
+        wget -q http://apt.puppetlabs.com/$DEB
+        sudo dpkg -i $DEB
+    fi
+    sudo apt-get update
+    sudo apt-get install --yes puppet
 
-#http://blog.doismellburning.co.uk/2013/01/19/upgrading-puppet-in-vagrant-boxes/
-apt-get install --yes lsb-release
-DISTRIB_CODENAME=$(lsb_release --codename --short)
-DEB="puppetlabs-release-${DISTRIB_CODENAME}.deb"
-# Assume that this file's existence means we have the Puppet Labs repo added
-DEB_PROVIDES="/etc/apt/sources.list.d/puppetlabs.list" 
-
-if [ ! -e $DEB_PROVIDES ]
-then
-    # Print statement useful for debugging, but automated runs of this will interpret any output as an error
-    # print "Could not find $DEB_PROVIDES - fetching and installing $DEB"
-    wget -q http://apt.puppetlabs.com/$DEB
-    sudo dpkg -i $DEB
+    # Note that we've been here
+    mkdir -p tmp
+    touch tmp/puppet-repo-added
 fi
-sudo apt-get update
-sudo apt-get install --yes puppet
-
+    
 echo "Puppet version: $(puppet --version) installed!"
+
+
